+++ src/autoskillit/recipe/rules/__init__.py
@@ -1,0 +1,1 @@
+"""Recipe semantic rule modules — imported for side-effect registration."""
