+++ src/autoskillit/recipe/rules/rules_campaign.py
