+++ src/autoskillit/recipe/rules/rules_isolation.py
