+++ src/autoskillit/recipe/rules/rules_temp_path.py
